+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Date format: DD/MM/YYYY
 
+## 03/03/2023
+
+🚀 **Additions**
+- Collection models are now exposed in the Swagger API documentation.
+
 ## 25/02/2023
 
 🚨 **Breaking changes**